requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==2.1.5
pycparser==2.22
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.1.2
Flask-Migrate==4.0.7